玩家血条已经完成

+++ momoka/include/core/tools/Gui.h
@@ -6,7 +6,7 @@ class Gui {
 	ID2D1Bitmap* enemyHPBar;
 public:
 	void DrawDebugText(std::string text);
-	void DrawHeroHp(int currentValue, int maxValue, int posX, int posY);
+	void DrawHeroHp(float currentValue, float maxValue, int posX, int posY);
 	void DrawEnemyHp(int currentValue, int maxValue, int posX, int posY);
 
 	Gui();
